Implement `GivenWallet`.
